fix: 작성자 외 스터디 삭제 금지 조정

+++ server/src/main/java/main_project/udongs/study/service/StudyService.java
@@ -82,9 +82,11 @@ public class StudyService {
 
     //스터디 삭제
     @Transactional
-    public void deleteStudy(Long studyId) {
+    public void deleteStudy(Long studyId, Long memberId) {
         Study findStudy = findVerifiedStudy(studyId);
-        studyRepository.delete(findStudy);
+        if(findStudy.getMember().getMemberId() == memberId) {
+            studyRepository.delete(findStudy);
+        } else throw new BusinessLogicException(ExceptionCode.UNAUTHORIZED_ACTION);
     }
